Changed the instance Size to support floating-point type of # of cores (ex, 6.5 cores). Propogated the change to template builder and related ComputeService usages.

+++ compute/src/main/java/org/jclouds/compute/domain/Size.java
@@ -40,7 +40,7 @@ public interface Size extends ComputeMetadata {
    /**
     * Amount of virtual or physical cores provided
     */
-   int getCores();
+   double getCores();
 
    /**
     * Amount of RAM provided in MB (256M, 1740)
